moving the feature vector from per edge base to per node base
Every node now has a feature vector and an array of (neighbor ids) instead of an array of edges (containing ids and features). This reduces the memory consumption but is much slower. It is a good comparison for the ReadOnlyGraph.

+++ benchmark/include/stopwatch.h
@@ -10,7 +10,7 @@ public:
         time_begin = std::chrono::steady_clock::now();
     }
 
-    float getElapsedTimeMicro()
+    long long getElapsedTimeMicro()
     {
         std::chrono::steady_clock::time_point time_end = std::chrono::steady_clock::now();
         return (std::chrono::duration_cast<std::chrono::microseconds>(time_end - time_begin).count());
